requirements.txt
# Scraping et collecte de données
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1

# AWS
boto3==1.33.6
botocore==1.33.6

# Base de données
psycopg2-binary==2.9.9
pandas==2.1.3
sqlalchemy==2.0.23

# Traitement des données
nltk==3.8.1
spacy==3.7.2
numpy==1.26.2
scipy==1.11.4

# Visualisation
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.2
plotly==5.18.0

# Utilitaires
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2
pytz==2023.3.post1
python-dateutil==2.8.2
six==1.16.0
tzdata==2023.3

# Développement
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.26.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0

